requirements.txt
tensorflow-gpu==2.10.0
PyYAML==5.4.1